Splits main package.
Moves code into internal packages to reduce the public API surface.
